pyproject.toml
[project]
name = "knowledgebase"
version = "0.1.0"
description = "KnowledgeBase MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "mcp>=1.8.0",
    "httpx>=0.28.0",
    "pymilvus>=2.6.0",
    "pypdf>=5.4.0",
    "python-multipart>=0.0.20",
    "python-docx>=1.1.2",
    "pydantic>=2.11.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.40",
]

[project.scripts]
knowledgebase = "knowledgebase.mcp.server:run"
knowledgebase-worker = "knowledgebase.worker.import_worker:run"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
